requirements.txt
torch~=2.0.1+cu118
pillow~=9.2.0
torchvision~=0.15.2+cu118
tqdm~=4.64.1
numpy~=1.24.4
ftfy~=6.1.1
regex~=2023.5.5
scipy~=1.10.1
h5py~=3.7.0
matplotlib~=3.4.3